move errors to separate package
