+++ pkg/cmd/pulumi/newcmd/config.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/pulumi/pulumi/pkg/v3/backend"
 	"github.com/pulumi/pulumi/pkg/v3/backend/display"
+	backendErrors "github.com/pulumi/pulumi/pkg/v3/backend/errors"
 	cmdConfig "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/config"
 	cmdStack "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/stack"
 	"github.com/pulumi/pulumi/pkg/v3/resource/deploy"
@@ -50,7 +51,7 @@ func HandleConfig(
 ) error {
 	// Get the existing config. stackConfig will be nil if there wasn't a previous deployment.
 	stackConfig, err := backend.GetLatestConfiguration(ctx, s)
-	if err != nil && err != backend.ErrNoPreviousDeployment {
+	if err != nil && err != backendErrors.ErrNoPreviousDeployment {
 		return err
 	}
